fix(db): run the right migration only for volumes
